requirements.txt
selenium==4.21.0
undetected-chromedriver==3.5.5
pdfminer.six
python-dotenv==1.0.1